refactor(app): extract lifecycle/quit/session methods (phase 6h)
Move should_quit, should_detach, clear_detach, set_session_mode,
is_session_mode, set_software_cursor_only, set_session_name,
session_name, queue_escape_sequences, take_pending_escape_sequences,
take_pending_clipboard, should_restart, take_restart_dir,
request_full_redraw, take_full_redraw_request, request_restart,
theme, is_settings_open, quit, count_modified_buffers_needing_prompt,
focus_gained, resize out of mod.rs into a new app/lifecycle.rs sibling.
~248 lines.

Editor lifecycle, session management, terminal escape-sequence and
clipboard piping, and the focus/resize hooks. None of these are
central editor state — they're orchestrators that coordinate session
state with the rest of the system.

mod.rs goes from 3289 -> 3042 lines (-247). 2026 tests still pass.
